Save/restore date start/end, salary in double format

+++ src/com/TheJobCoach/webapp/userpage/shared/UserOpportunity.java
@@ -43,7 +43,7 @@ public class UserOpportunity implements Serializable {
 	public String contractType;
 
 	/** How much you expect to earn */
-	public int salary;
+	public double salary;
 
 	/** When the job starts. Set to 1/1/1900 if not set */	
 	public Date startDate;
@@ -96,7 +96,7 @@ public class UserOpportunity implements Serializable {
 
 	public UserOpportunity(String iD, Date firstSeen, Date lastUpdate,
 			String title, String description, String companyId,
-			String contractType, int salary, Date startDate, Date endDate,
+			String contractType, double salary, Date startDate, Date endDate,
 			boolean systemSource, String source, String url, String location,
 			ApplicationStatus status, String note) 
 	{
